refactor: make scene class follow DRY principles.
previously, we had to write a render function for each scene which was derived from scene class although the content of the render function was almost similar for every scene fixed that by making render a virtual function and adding implementation in base class + fixed outline issues in test scene which were due to a dirty stencil buffer state i.e. stencil buffer state was not reset after drawing the outline which caused unexpected behaviour.